Refactor code for generating random ip (#3080)
Summary:
Fixes # (issue)
This PR moves route_through_default_routes and generate_ip_through_default_route from test_bgpmon.py to generators.py since same logic is needed for dualtor test cases.
